refactor main project to use new folders structure, add BoolToVisibilityConverter to application resources to use without referencing in user-controls
