+++ AmSpaceTools/App.xaml
@@ -2,6 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:AmSpaceTools"
+             xmlns:converters="clr-namespace:AmSpaceTools.Infrastructure.Converters"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
         <ResourceDictionary>
@@ -18,6 +19,7 @@
                     <SolidColorBrush x:Key="PrimaryError" Color="#AB0743"/> <!--AmRest red-->
                 </ResourceDictionary>
             </ResourceDictionary.MergedDictionaries>
+            <converters:BoolToVisibilityConverter x:Key="BoolToVisibilityConverter"/>
         </ResourceDictionary>
     </Application.Resources>
 </Application>
