+++ AmSpaceTools/Infrastructure/Extensions/StringExtensions.cs
@@ -5,7 +5,7 @@ using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AmSpaceTools.Infrastructure
+namespace AmSpaceTools.Infrastructure.Extensions
 {
     public static class StringExtensions
     {
