+++ AmSpaceTools/Infrastructure/Extensions/TreeItem.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AmSpaceTools.Infrastructure
+namespace AmSpaceTools.Infrastructure.Extensions
 {
     public class TreeItem<T>
     {
