+++ AmSpaceTools/ViewModels/BaseViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using AmSpaceTools.Infrastructure;
+using AmSpaceTools.Infrastructure.Extensions;
 using System;
 using System.Collections;
 using System.Collections.Generic;
