+++ AmSpaceTools/ViewModels/JobMapUploadViewModel.cs
@@ -2,6 +2,7 @@
 using AmSpaceModels.JobMap;
 using AmSpaceModels.Organization;
 using AmSpaceTools.Infrastructure;
+using AmSpaceTools.Infrastructure.Extensions;
 using AmSpaceTools.Properties;
 using ExcelWorker;
 using ExcelWorker.Models;
