+++ AmSpaceTools/ViewModels/LoginViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using AmSpaceClient;
 using AmSpaceModels;
 using AmSpaceTools.Infrastructure;
+using AmSpaceTools.Infrastructure.Providers;
 using AmSpaceTools.Views;
 using StructureMap.Pipeline;
 using System;
