+++ AmSpaceTools/ViewModels/OrgStructureViewModel.cs
@@ -2,6 +2,7 @@
 using AmSpaceModels.Organization;
 using AmSpaceModels.Sap;
 using AmSpaceTools.Infrastructure;
+using AmSpaceTools.Infrastructure.Extensions;
 using AmSpaceTools.Views;
 using AutoMapper;
 using ExcelWorker;
