+++ AmSpaceTools/ViewModels/SearchPeopleViewModel.cs
@@ -2,6 +2,7 @@
 using AmSpaceModels.Enums;
 using AmSpaceModels.Organization;
 using AmSpaceTools.Infrastructure;
+using AmSpaceTools.Infrastructure.Extensions;
 using ExcelWorker.Models;
 using MaterialDesignThemes.Wpf;
 using System;
