+++ AmSpaceTools/Views/ChangePassword.xaml
@@ -8,9 +8,6 @@
              xmlns:infrastructure="clr-namespace:AmSpaceTools.Infrastructure"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
-    <UserControl.Resources>
-        <infrastructure:BoolToVisibilityConverter x:Key="BoolToVisibilityConverter"/>
-    </UserControl.Resources>
     <StackPanel Margin="10" Width="300">
         <TextBlock
                 Margin="0 5 5 2"
