+++ AmSpaceTools/Views/IdpTranslationsPreviewer.xaml
@@ -12,7 +12,6 @@
              mc:Ignorable="d" 
              d:DesignHeight="300" d:DesignWidth="300">
     <UserControl.Resources>
-        <infrastructure:BoolToVisibilityConverter x:Key="BoolToVisibilityConverter"/>
         <ObjectDataProvider x:Key="ColumnActionFromEnum"
                             MethodName="GetValues" 
                             ObjectType="{x:Type sys:Enum}">
