+++ AmSpaceTools/Views/PeopleUpload.xaml
@@ -6,13 +6,9 @@
              xmlns:modelsSap="clr-namespace:AmSpaceModels.Sap;assembly=AmSpaceModels"
              xmlns:sys="clr-namespace:System;assembly=mscorlib"
              xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
-             xmlns:infrastructure="clr-namespace:AmSpaceTools.Infrastructure"
              x:Class="AmSpaceTools.Views.PeopleUpload"
              mc:Ignorable="d" 
              d:DesignHeight="300" d:DesignWidth="300">
-    <UserControl.Resources>
-        <infrastructure:BoolToVisibilityConverter x:Key="BoolToVisibilityConverter"/>
-    </UserControl.Resources>
     <materialDesign:DialogHost Identifier="ControlDialog">
         <Grid IsEnabled="{Binding UIEnabled}">
             <Grid.ColumnDefinitions>
